#8 refactor : inner class를 리팩토링 한다
