+++ core/src/main/java/kusitms/duduk/core/newsletter/port/input/RetrieveNewsLetterQuery.java
@@ -1,0 +1,8 @@
+package kusitms.duduk.core.newsletter.port.input;
+
+import kusitms.duduk.core.newsletter.dto.request.RetrieveNewsLetterRequest;
+import kusitms.duduk.core.newsletter.dto.response.NewsLetterResponse;
+
+public interface RetrieveNewsLetterQuery {
+    NewsLetterResponse retrieve(RetrieveNewsLetterRequest request);
+}
